Move get_missing_fonts() and warn_about_missing_fonts() from PdfPage to PdfTextPage where they logically belong. Remove broad exception catching to let errors propagate naturally. Add comprehensive test coverage for the new methods.
